// src/app/shared/models/date-picker.models.ts
export class datePickerConfig {
  dateInputFormat: string;
  containerClass: string;
  selectFromOtherMonth: boolean;
  isAnimated: boolean;
  adaptivePosition: boolean;
  showWeekNumbers: boolean;
  returnFocusToInput: boolean;
  isDisabled: boolean;
  clearButtonLabel: string;
  clearPosition: string;
  showClearButton: boolean;
}